+++ spec/fabricators/invoice_fabricator.rb
@@ -2,4 +2,7 @@
 
 Fabricator(:invoice) do
   payment_vendor 'paypal'
+  accounts do
+    [{ email: Faker::Internet.email, display_name: Faker::Company.name }]
+  end
 end
